Added more methods to the protocol source.
Added a method to iterate through an array of protocols to check each
for them once a packet is received. Also made a method to generate a
Unknown Protocol whenever the protocol doesn't correspond to one on the
list. It is a error value if you must but it could also be a indication
that the protocol isn't on the list of unsecure ones for example.