Added info extraction from rule set
Added some mechanisms to make it possible to account for the rule
quantification and counters in order to provide the user with
notifications or even a report with all the quantities. If the rule set
changes then the counters are all reseted and the data within the DB is
also droped because it is of no use, the rules are not the same.